解决项目变更查看详情页数据加载错误的问题，解决责任人补填当前一周工时日期校验逻辑错误的问题，去掉系统消息

+++ alm/src/main/java/com/dili/alm/dao/WorkDayMapper.java
@@ -18,5 +18,5 @@ public interface WorkDayMapper extends MyMapper<WorkDay> {
 
 	Integer deteleWorkDaysByYear(@Param("year") String year);
 
-	Integer countByStartAndEndDate(@Param("startDate") Date startDate, @Param("endDate") Date endDate);
+	Long countByStartAndEndDate(@Param("startDate") Date startDate, @Param("endDate") Date endDate);
 }
